+++ infrastructure/__init__.py
@@ -4,6 +4,7 @@ from .eventbus import InMemoryEventBus
 from .graph_builder import InMemoryGraphBuilder
 from .config_loader import ConfigLoader
 from .snapshot_store import SnapshotStore
+from .plugin_loader import PluginLoader
 
 __all__ = [
     "DependencyContainer",
@@ -11,4 +12,5 @@ __all__ = [
     "InMemoryGraphBuilder",
     "ConfigLoader",
     "SnapshotStore",
+    "PluginLoader",
 ]